Restructured some shit
- removed some hypixel remnants

-> 1.24.3.2
