+++ src/main/java/de/klassenserver7b/k7bot/manage/PropertiesManager.java
@@ -65,7 +65,6 @@ public class PropertiesManager {
         String token = prop.getProperty("token");
 
         if (token != null && !token.isBlank()) {
-
             return token.split("\\.").length == 3;
         }
