+++ src/main/java/de/klassenserver7b/k7bot/music/commands/generic/GenericChartsCommand.java
@@ -3,10 +3,10 @@
  */
 package de.klassenserver7b.k7bot.music.commands.generic;
 
-import de.klassenserver7b.k7bot.util.Cell;
+import de.klassenserver7b.k7bot.util.tablemessage.Cell;
 import de.klassenserver7b.k7bot.util.EmbedUtils;
 import de.klassenserver7b.k7bot.util.GenericMessageSendHandler;
-import de.klassenserver7b.k7bot.util.TableMessage;
+import de.klassenserver7b.k7bot.util.tablemessage.TableMessage;
 import net.dv8tion.jda.api.EmbedBuilder;
 
 import java.awt.*;
