+++ src/main/java/de/klassenserver7b/k7bot/music/commands/slash/PlaySlashCommand.java
@@ -8,7 +8,7 @@ import de.klassenserver7b.k7bot.music.commands.generic.GenericPlayCommand;
 import de.klassenserver7b.k7bot.music.lavaplayer.AudioLoadResult;
 import de.klassenserver7b.k7bot.music.lavaplayer.MusicController;
 import de.klassenserver7b.k7bot.music.utilities.AudioLoadOption;
-import de.klassenserver7b.k7bot.util.SupportedPlayQueries;
+import de.klassenserver7b.k7bot.music.SupportedPlayQueries;
 import net.dv8tion.jda.api.interactions.commands.Command.Choice;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
 import net.dv8tion.jda.api.interactions.commands.build.OptionData;
