+++ src/main/java/de/klassenserver7b/k7bot/util/customapis/VPlan_MANOS_OLD.java
@@ -7,9 +7,9 @@ import com.google.gson.JsonParser;
 import de.klassenserver7b.k7bot.Klassenserver7bbot;
 import de.klassenserver7b.k7bot.sql.LiteSQL;
 import de.klassenserver7b.k7bot.subscriptions.types.SubscriptionTarget;
-import de.klassenserver7b.k7bot.util.Cell;
+import de.klassenserver7b.k7bot.util.tablemessage.Cell;
 import de.klassenserver7b.k7bot.util.EmbedUtils;
-import de.klassenserver7b.k7bot.util.TableMessage;
+import de.klassenserver7b.k7bot.util.tablemessage.TableMessage;
 import de.klassenserver7b.k7bot.util.TeacherDB;
 import net.dv8tion.jda.annotations.DeprecatedSince;
 import net.dv8tion.jda.api.EmbedBuilder;
